added header file containing class defs
