+++ CUGeneration/CUGenerationPass.cpp
@@ -275,304 +275,304 @@ using namespace dputil;
 /*****************************   DiscoPoP Functions
  * ***********************************/
 
-void CUGeneration::getFunctionReturnLines(Region *TopRegion, Node *root) {
-  int lid = 0;
-  for (Region::block_iterator bb = TopRegion->block_begin();
-       bb != TopRegion->block_end(); ++bb) {
-    for (BasicBlock::iterator instruction = (*bb)->begin();
-         instruction != (*bb)->end(); ++instruction) {
-      if (isa<StoreInst>(instruction)) {
-        string varName = determineVariableName(&*instruction);
-        size_t pos = varName.find("retval");
-        if (pos != varName.npos) {
-          lid = getLID(&*instruction, fileID);
-          // errs() << "varName: " << varName << " " << dputil::decodeLID(lid)
-          // << "\n";
-          if (lid > 0)
-            root->returnLines.insert(lid);
-        }
-      }
-    }
-  }
-}
+// void CUGeneration::getFunctionReturnLines(Region *TopRegion, Node *root) {
+//   int lid = 0;
+//   for (Region::block_iterator bb = TopRegion->block_begin();
+//        bb != TopRegion->block_end(); ++bb) {
+//     for (BasicBlock::iterator instruction = (*bb)->begin();
+//          instruction != (*bb)->end(); ++instruction) {
+//       if (isa<StoreInst>(instruction)) {
+//         string varName = determineVariableName(&*instruction);
+//         size_t pos = varName.find("retval");
+//         if (pos != varName.npos) {
+//           lid = getLID(&*instruction, fileID);
+//           // errs() << "varName: " << varName << " " << dputil::decodeLID(lid)
+//           // << "\n";
+//           if (lid > 0)
+//             root->returnLines.insert(lid);
+//         }
+//       }
+//     }
+//   }
+// }
 
-string CUGeneration::determineVariableDefLine(Instruction *I) {
-  string varDefLine{"LineNotFound"};
+// string CUGeneration::determineVariableDefLine(Instruction *I) {
+//   string varDefLine{"LineNotFound"};
 
-  string varName = determineVariableName(&*I);
-  // varName = refineVarName(varName);
-  varName = (varName.find(".addr") == varName.npos)
-                ? varName
-                : varName.erase(varName.find(".addr"), 5);
-  // varName.erase(varName.find(".addr"), 5);
-  // size_t pos = varName.find(".addr");
-  // if (pos != varName.npos)
-  //     varName.erase(varName.find(".addr"), 5);
+//   string varName = determineVariableName(&*I);
+//   // varName = refineVarName(varName);
+//   varName = (varName.find(".addr") == varName.npos)
+//                 ? varName
+//                 : varName.erase(varName.find(".addr"), 5);
+//   // varName.erase(varName.find(".addr"), 5);
+//   // size_t pos = varName.find(".addr");
+//   // if (pos != varName.npos)
+//   //     varName.erase(varName.find(".addr"), 5);
 
-  string varType = determineVariableType(&*I);
+//   string varType = determineVariableType(&*I);
 
-  if (programGlobalVariablesSet.count(varName)) {
-    varDefLine = "GlobalVar";
-    // TODO: Find definition line of global variables
-  }
+//   if (programGlobalVariablesSet.count(varName)) {
+//     varDefLine = "GlobalVar";
+//     // TODO: Find definition line of global variables
+//   }
 
-  // Start from the beginning of a function and look for the variable
-  Function *F = I->getFunction();
-  for (Function::iterator FI = F->begin(), FE = F->end(); FI != FE; ++FI) {
-    BasicBlock &BB = *FI;
-    for (BasicBlock::iterator BI = BB.begin(), E = BB.end(); BI != E; ++BI) {
-      if (DbgDeclareInst *DI = dyn_cast<DbgDeclareInst>(BI)) {
+//   // Start from the beginning of a function and look for the variable
+//   Function *F = I->getFunction();
+//   for (Function::iterator FI = F->begin(), FE = F->end(); FI != FE; ++FI) {
+//     BasicBlock &BB = *FI;
+//     for (BasicBlock::iterator BI = BB.begin(), E = BB.end(); BI != E; ++BI) {
+//       if (DbgDeclareInst *DI = dyn_cast<DbgDeclareInst>(BI)) {
 
-        if (auto *N = dyn_cast<MDNode>(DI->getVariable())) {
-          if (auto *DV = dyn_cast<DILocalVariable>(N)) {
-            if (varType.find("ARRAY") != string::npos ||
-                varType.find("STRUCT") != string::npos) {
-              if (DV->getName() == varName) {
-                varDefLine = to_string(fileID) + ":" + to_string(DV->getLine());
-                break;
-              }
-            } else {
-              string vn = "----";
-              AllocaInst *AI = dyn_cast_or_null<AllocaInst>(DI->getAddress());
-              if (AI) {
-                for (User *U : AI->users()) {
-                  if (StoreInst *SI = dyn_cast<StoreInst>(U)) {
-                    vn = determineVariableName(&*SI);
-                    break;
-                  } else if (LoadInst *LI = dyn_cast<LoadInst>(U)) {
-                    vn = determineVariableName(&*LI);
-                    break;
-                  }
-                }
-                if (vn == varName) {
-                  varDefLine =
-                      to_string(fileID) + ":" + to_string(DV->getLine());
-                  break;
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-  }
-  return varDefLine;
-}
+//         if (auto *N = dyn_cast<MDNode>(DI->getVariable())) {
+//           if (auto *DV = dyn_cast<DILocalVariable>(N)) {
+//             if (varType.find("ARRAY") != string::npos ||
+//                 varType.find("STRUCT") != string::npos) {
+//               if (DV->getName() == varName) {
+//                 varDefLine = to_string(fileID) + ":" + to_string(DV->getLine());
+//                 break;
+//               }
+//             } else {
+//               string vn = "----";
+//               AllocaInst *AI = dyn_cast_or_null<AllocaInst>(DI->getAddress());
+//               if (AI) {
+//                 for (User *U : AI->users()) {
+//                   if (StoreInst *SI = dyn_cast<StoreInst>(U)) {
+//                     vn = determineVariableName(&*SI);
+//                     break;
+//                   } else if (LoadInst *LI = dyn_cast<LoadInst>(U)) {
+//                     vn = determineVariableName(&*LI);
+//                     break;
+//                   }
+//                 }
+//                 if (vn == varName) {
+//                   varDefLine =
+//                       to_string(fileID) + ":" + to_string(DV->getLine());
+//                   break;
+//                 }
+//               }
+//             }
+//           }
+//         }
+//       }
+//     }
+//   }
+//   return varDefLine;
+// }
 
-string CUGeneration::determineVariableType(Instruction *I) {
-  string s = "";
-  string type_str;
-  int index = isa<StoreInst>(I) ? 1 : 0;
-  raw_string_ostream rso(type_str);
-  (*((I->getOperand(index))->getType())).print(rso);
+// string CUGeneration::determineVariableType(Instruction *I) {
+//   string s = "";
+//   string type_str;
+//   int index = isa<StoreInst>(I) ? 1 : 0;
+//   raw_string_ostream rso(type_str);
+//   (*((I->getOperand(index))->getType())).print(rso);
 
-  Value *operand = I->getOperand(index);
+//   Value *operand = I->getOperand(index);
 
-  if (operand->hasName()) {
-    if (isa<GetElementPtrInst>(*operand)) {
-      GetElementPtrInst *gep = cast<GetElementPtrInst>(operand);
-      Value *ptrOperand = gep->getPointerOperand();
-      PointerType *PTy = cast<PointerType>(ptrOperand->getType());
-      // we've found a struct/class
-      Type *structType = pointsToStruct(PTy);
-      if (structType && gep->getNumOperands() > 2) {
-        s = "STRUCT,";
-      }
-      // we've found an array
-      if (PTy->getElementType()->getTypeID() == Type::ArrayTyID) {
-        s = "ARRAY,";
-      }
-    }
-  }
+//   if (operand->hasName()) {
+//     if (isa<GetElementPtrInst>(*operand)) {
+//       GetElementPtrInst *gep = cast<GetElementPtrInst>(operand);
+//       Value *ptrOperand = gep->getPointerOperand();
+//       PointerType *PTy = cast<PointerType>(ptrOperand->getType());
+//       // we've found a struct/class
+//       Type *structType = pointsToStruct(PTy);
+//       if (structType && gep->getNumOperands() > 2) {
+//         s = "STRUCT,";
+//       }
+//       // we've found an array
+//       if (PTy->getElementType()->getTypeID() == Type::ArrayTyID) {
+//         s = "ARRAY,";
+//       }
+//     }
+//   }
 
-  s = s + rso.str();
-  return s;
-}
+//   s = s + rso.str();
+//   return s;
+// }
 
-string CUGeneration::determineVariableName(
-    Instruction *I, bool &isGlobalVariable /*=defaultIsGlobalVariableValue*/) {
+// string CUGeneration::determineVariableName(
+//     Instruction *I, bool &isGlobalVariable /*=defaultIsGlobalVariableValue*/) {
 
-  assert(I && "Instruction cannot be NULL \n");
-  int index = isa<StoreInst>(I) ? 1 : 0;
-  Value *operand = I->getOperand(index);
+//   assert(I && "Instruction cannot be NULL \n");
+//   int index = isa<StoreInst>(I) ? 1 : 0;
+//   Value *operand = I->getOperand(index);
 
-  IRBuilder<> builder(I);
+//   IRBuilder<> builder(I);
 
-  if (operand == NULL) {
-    return getOrInsertVarName("", builder);
-  }
+//   if (operand == NULL) {
+//     return getOrInsertVarName("", builder);
+//   }
 
-  if (operand->hasName()) {
-    //// we've found a global variable
-    if (isa<GlobalVariable>(*operand)) {
-      // MOHAMMAD ADDED THIS FOR CHECKING
-      isGlobalVariable = true;
-      return string(operand->getName());
-    }
-    if (isa<GetElementPtrInst>(*operand)) {
-      GetElementPtrInst *gep = cast<GetElementPtrInst>(operand);
-      Value *ptrOperand = gep->getPointerOperand();
-      PointerType *PTy = cast<PointerType>(ptrOperand->getType());
+//   if (operand->hasName()) {
+//     //// we've found a global variable
+//     if (isa<GlobalVariable>(*operand)) {
+//       // MOHAMMAD ADDED THIS FOR CHECKING
+//       isGlobalVariable = true;
+//       return string(operand->getName());
+//     }
+//     if (isa<GetElementPtrInst>(*operand)) {
+//       GetElementPtrInst *gep = cast<GetElementPtrInst>(operand);
+//       Value *ptrOperand = gep->getPointerOperand();
+//       PointerType *PTy = cast<PointerType>(ptrOperand->getType());
 
-      // we've found a struct/class
-      Type *structType = pointsToStruct(PTy);
-      if (structType && gep->getNumOperands() > 2) {
-        Value *constValue = gep->getOperand(2);
-        if (constValue && isa<ConstantInt>(*constValue)) {
-          ConstantInt *idxPtr = cast<ConstantInt>(gep->getOperand(2));
-          uint64_t memberIdx = *(idxPtr->getValue().getRawData());
-          if (!(cast<StructType>(structType))->isLiteral()) {
-            string strName(structType->getStructName().data());
-            map<string, MDNode *>::iterator it = Structs.find(strName);
-            if (it != Structs.end()) {
-              std::string ret =
-                  findStructMemberName(it->second, memberIdx, builder);
-              if (ret.size() > 0)
-                return ret;
-              else
-                return getOrInsertVarName("", builder);
-              // return ret;
-            }
-          }
-        }
-      }
+//       // we've found a struct/class
+//       Type *structType = pointsToStruct(PTy);
+//       if (structType && gep->getNumOperands() > 2) {
+//         Value *constValue = gep->getOperand(2);
+//         if (constValue && isa<ConstantInt>(*constValue)) {
+//           ConstantInt *idxPtr = cast<ConstantInt>(gep->getOperand(2));
+//           uint64_t memberIdx = *(idxPtr->getValue().getRawData());
+//           if (!(cast<StructType>(structType))->isLiteral()) {
+//             string strName(structType->getStructName().data());
+//             map<string, MDNode *>::iterator it = Structs.find(strName);
+//             if (it != Structs.end()) {
+//               std::string ret =
+//                   findStructMemberName(it->second, memberIdx, builder);
+//               if (ret.size() > 0)
+//                 return ret;
+//               else
+//                 return getOrInsertVarName("", builder);
+//               // return ret;
+//             }
+//           }
+//         }
+//       }
 
-      // we've found an array
-      if (PTy->getElementType()->getTypeID() == Type::ArrayTyID &&
-          isa<GetElementPtrInst>(*ptrOperand)) {
-        return determineVariableName((Instruction *)ptrOperand,
-                                     isGlobalVariable);
-      }
-      return determineVariableName((Instruction *)gep, isGlobalVariable);
-    }
-    return string(operand->getName().data());
-    // return getOrInsertVarName(string(operand->getName().data()), builder);
-  }
+//       // we've found an array
+//       if (PTy->getElementType()->getTypeID() == Type::ArrayTyID &&
+//           isa<GetElementPtrInst>(*ptrOperand)) {
+//         return determineVariableName((Instruction *)ptrOperand,
+//                                      isGlobalVariable);
+//       }
+//       return determineVariableName((Instruction *)gep, isGlobalVariable);
+//     }
+//     return string(operand->getName().data());
+//     // return getOrInsertVarName(string(operand->getName().data()), builder);
+//   }
 
-  if (isa<LoadInst>(*operand) || isa<StoreInst>(*operand)) {
-    return determineVariableName((Instruction *)(operand), isGlobalVariable);
-  }
-  // if we cannot determine the name, then return *
-  return ""; // getOrInsertVarName("*", builder);
-}
+//   if (isa<LoadInst>(*operand) || isa<StoreInst>(*operand)) {
+//     return determineVariableName((Instruction *)(operand), isGlobalVariable);
+//   }
+//   // if we cannot determine the name, then return *
+//   return ""; // getOrInsertVarName("*", builder);
+// }
 
-Type *CUGeneration::pointsToStruct(PointerType *PTy) {
-  assert(PTy);
-  Type *structType = PTy;
-  if (PTy->getTypeID() == Type::PointerTyID) {
-    while (structType->getTypeID() == Type::PointerTyID) {
-      structType = cast<PointerType>(structType)->getElementType();
-    }
-  }
-  return structType->getTypeID() == Type::StructTyID ? structType : NULL;
-}
+// Type *CUGeneration::pointsToStruct(PointerType *PTy) {
+//   assert(PTy);
+//   Type *structType = PTy;
+//   if (PTy->getTypeID() == Type::PointerTyID) {
+//     while (structType->getTypeID() == Type::PointerTyID) {
+//       structType = cast<PointerType>(structType)->getElementType();
+//     }
+//   }
+//   return structType->getTypeID() == Type::StructTyID ? structType : NULL;
+// }
 
-string CUGeneration::getOrInsertVarName(string varName, IRBuilder<> &builder) {
-  Value *valName = NULL;
-  std::string vName = varName;
-  map<string, Value *>::iterator pair = VarNames.find(varName);
-  if (pair == VarNames.end()) {
-    valName = builder.CreateGlobalStringPtr(StringRef(varName.c_str()), ".str");
+// string CUGeneration::getOrInsertVarName(string varName, IRBuilder<> &builder) {
+//   Value *valName = NULL;
+//   std::string vName = varName;
+//   map<string, Value *>::iterator pair = VarNames.find(varName);
+//   if (pair == VarNames.end()) {
+//     valName = builder.CreateGlobalStringPtr(StringRef(varName.c_str()), ".str");
 
-    VarNames[varName] = valName;
-  } else {
-    vName = pair->first;
-  }
+//     VarNames[varName] = valName;
+//   } else {
+//     vName = pair->first;
+//   }
 
-  return vName;
-}
+//   return vName;
+// }
 
-string CUGeneration::findStructMemberName(MDNode *structNode, unsigned idx,
-                                          IRBuilder<> &builder) {
-  assert(structNode);
-  assert(structNode->getOperand(10));
-  MDNode *memberListNodes = cast<MDNode>(structNode->getOperand(10));
-  if (idx < memberListNodes->getNumOperands()) {
-    assert(memberListNodes->getOperand(idx));
-    MDNode *member = cast<MDNode>(memberListNodes->getOperand(idx));
-    if (member->getOperand(3)) {
-      // getOrInsertVarName(string(member->getOperand(3)->getName().data()),
-      // builder); return string(member->getOperand(3)->getName().data());
-      getOrInsertVarName(dyn_cast<MDString>(member->getOperand(3))->getString(),
-                         builder);
-      return dyn_cast<MDString>(member->getOperand(3))->getString();
-    }
-  }
-  return NULL;
-}
+// string CUGeneration::findStructMemberName(MDNode *structNode, unsigned idx,
+//                                           IRBuilder<> &builder) {
+//   assert(structNode);
+//   assert(structNode->getOperand(10));
+//   MDNode *memberListNodes = cast<MDNode>(structNode->getOperand(10));
+//   if (idx < memberListNodes->getNumOperands()) {
+//     assert(memberListNodes->getOperand(idx));
+//     MDNode *member = cast<MDNode>(memberListNodes->getOperand(idx));
+//     if (member->getOperand(3)) {
+//       // getOrInsertVarName(string(member->getOperand(3)->getName().data()),
+//       // builder); return string(member->getOperand(3)->getName().data());
+//       getOrInsertVarName(dyn_cast<MDString>(member->getOperand(3))->getString(),
+//                          builder);
+//       return dyn_cast<MDString>(member->getOperand(3))->getString();
+//     }
+//   }
+//   return NULL;
+// }
 /*********************************** End of DiscoPoP Functions
  * ***********************************/
 
 /*********************************** Output functions
  * ********************************************/
-string CUGeneration::xmlEscape(string data) {
-  string::size_type pos = 0;
-  for (;;) {
-    pos = data.find_first_of("\"&<>", pos);
-    if (pos == string::npos)
-      break;
-    string replacement;
-    switch (data[pos]) {
-    case '\"':
-      replacement = "&quot;";
-      break;
-    case '&':
-      replacement = "&amp;";
-      break;
-    case '<':
-      replacement = "&lt;";
-      break;
-    case '>':
-      replacement = "&gt;";
-      break;
-    default:;
-    }
-    data.replace(pos, 1, replacement);
-    pos += replacement.size();
-  };
-  return data;
-}
+// string CUGeneration::xmlEscape(string data) {
+//   string::size_type pos = 0;
+//   for (;;) {
+//     pos = data.find_first_of("\"&<>", pos);
+//     if (pos == string::npos)
+//       break;
+//     string replacement;
+//     switch (data[pos]) {
+//     case '\"':
+//       replacement = "&quot;";
+//       break;
+//     case '&':
+//       replacement = "&amp;";
+//       break;
+//     case '<':
+//       replacement = "&lt;";
+//       break;
+//     case '>':
+//       replacement = "&gt;";
+//       break;
+//     default:;
+//     }
+//     data.replace(pos, 1, replacement);
+//     pos += replacement.size();
+//   };
+//   return data;
+// }
 
-void CUGeneration::secureStream() {
-  outOriginalVariables = new std::ofstream();
-  outOriginalVariables->open("OriginalVariables.txt", std::ios_base::app);
+// void CUGeneration::secureStream() {
+//   outOriginalVariables = new std::ofstream();
+//   outOriginalVariables->open("OriginalVariables.txt", std::ios_base::app);
 
-  outCUs = new std::ofstream();
-  outCUs->open("Data.xml", std::ios_base::app);
+//   outCUs = new std::ofstream();
+//   outCUs->open("Data.xml", std::ios_base::app);
 
-  outCUIDCounter = new std::ofstream();
-  outCUIDCounter->open("DP_CUIDCounter.txt", std::ios_base::out);
-}
+//   outCUIDCounter = new std::ofstream();
+//   outCUIDCounter->open("DP_CUIDCounter.txt", std::ios_base::out);
+// }
 
-string CUGeneration::getLineNumbersString(set<int> LineNumbers) {
-  string line = "";
-  for (auto li : LineNumbers) {
-    std::string temp = ',' + dputil::decodeLID(li);
-    if (temp != ",*") {
-      if (line == "") {
-        line = dputil::decodeLID(li);
-      } else {
-        line = line + temp;
-      }
-    }
-  }
-  return line;
-}
+// string CUGeneration::getLineNumbersString(set<int> LineNumbers) {
+//   string line = "";
+//   for (auto li : LineNumbers) {
+//     std::string temp = ',' + dputil::decodeLID(li);
+//     if (temp != ",*") {
+//       if (line == "") {
+//         line = dputil::decodeLID(li);
+//       } else {
+//         line = line + temp;
+//       }
+//     }
+//   }
+//   return line;
+// }
 
-string CUGeneration::getChildrenNodesString(Node *root) {
-  string childrenIDs = "";
-  int i = 0;
-  std::for_each(root->childrenNodes.begin(), root->childrenNodes.end(),
-                [&](Node *node) {
-                  if (i == 0) {
-                    childrenIDs = node->ID;
-                    i++;
-                  } else {
-                    childrenIDs += "," + node->ID;
-                  }
-                });
-  return childrenIDs;
-}
+// string CUGeneration::getChildrenNodesString(Node *root) {
+//   string childrenIDs = "";
+//   int i = 0;
+//   std::for_each(root->childrenNodes.begin(), root->childrenNodes.end(),
+//                 [&](Node *node) {
+//                   if (i == 0) {
+//                     childrenIDs = node->ID;
+//                     i++;
+//                   } else {
+//                     childrenIDs += "," + node->ID;
+//                   }
+//                 });
+//   return childrenIDs;
+// }
 
 // string CUGeneration::getChildrenNodesString(Node *root)
 // {
@@ -605,15 +605,15 @@ string CUGeneration::getChildrenNodesString(Node *root) {
 //     }
 // }
 
-void CUGeneration::printData(Node *root) {
-  *outCUs << "<Nodes>" << endl << endl;
+// void CUGeneration::printData(Node *root) {
+//   *outCUs << "<Nodes>" << endl << endl;
 
-  printTree(root, true);
+//   printTree(root, true);
 
-  *outCUs << "</Nodes>" << endl << endl << endl;
+//   *outCUs << "</Nodes>" << endl << endl << endl;
 
-  closeOutputFiles();
-}
+//   closeOutputFiles();
+// }
 
 // void CUGeneration::printTree(Node *root, bool isRoot)
 // {
@@ -631,133 +631,133 @@ void CUGeneration::printData(Node *root) {
 //     }
 // }
 
-void CUGeneration::printTree(Node *root, bool isRoot) {
-  printNode(root, isRoot);
+// void CUGeneration::printTree(Node *root, bool isRoot) {
+//   printNode(root, isRoot);
 
-  std::for_each(root->childrenNodes.begin(), root->childrenNodes.end(),
-                [&](Node *node) {
-                  if (node->type == nodeTypes::func)
-                    isRoot = false;
-                  printTree(node, isRoot);
-                });
-}
+//   std::for_each(root->childrenNodes.begin(), root->childrenNodes.end(),
+//                 [&](Node *node) {
+//                   if (node->type == nodeTypes::func)
+//                     isRoot = false;
+//                   printTree(node, isRoot);
+//                 });
+// }
 
-void CUGeneration::printNode(Node *root, bool isRoot) {
-  if (root->name.find("llvm")) {
-    // if(dputil::decodeLID(root->endLine) == "1:560")
-    // errs() << "=-=-=-=-=-=-=-=-=-=-=-=- " <<
-    // dputil::decodeLID(root->startLine) << "\n";
-    string start = "";
-    if (root->type == nodeTypes::loop) {
-      start = loopStartLines[root->ID];
-    } else {
-      start = dputil::decodeLID(root->startLine);
-    }
-    *outCUs << "\t<Node"
-            << " id=\"" << xmlEscape(root->ID) << "\""
-            << " type=\"" << root->type << "\""
-            << " name=\"" << xmlEscape(root->name) << "\""
-            << " startsAtLine = \"" << start << "\""
-            << " endsAtLine = \"" << dputil::decodeLID(root->endLine) << "\""
-            << ">" << endl;
-    *outCUs << "\t\t<childrenNodes>" << getChildrenNodesString(root)
-            << "</childrenNodes>" << endl;
-    if (root->type == nodeTypes::func || root->type == nodeTypes::dummy) {
-      *outCUs << "\t\t<funcArguments>" << endl;
-      for (auto ai : root->argumentsList) {
-        *outCUs << "\t\t\t<arg type=\"" << xmlEscape(ai.type) << "\""
-                << " defLine=\"" << xmlEscape(ai.defLine) << "\">"
-                << xmlEscape(ai.name) << "</arg>" << endl;
-      }
-      *outCUs << "\t\t</funcArguments>" << endl;
+// void CUGeneration::printNode(Node *root, bool isRoot) {
+//   if (root->name.find("llvm")) {
+//     // if(dputil::decodeLID(root->endLine) == "1:560")
+//     // errs() << "=-=-=-=-=-=-=-=-=-=-=-=- " <<
+//     // dputil::decodeLID(root->startLine) << "\n";
+//     string start = "";
+//     if (root->type == nodeTypes::loop) {
+//       start = loopStartLines[root->ID];
+//     } else {
+//       start = dputil::decodeLID(root->startLine);
+//     }
+//     *outCUs << "\t<Node"
+//             << " id=\"" << xmlEscape(root->ID) << "\""
+//             << " type=\"" << root->type << "\""
+//             << " name=\"" << xmlEscape(root->name) << "\""
+//             << " startsAtLine = \"" << start << "\""
+//             << " endsAtLine = \"" << dputil::decodeLID(root->endLine) << "\""
+//             << ">" << endl;
+//     *outCUs << "\t\t<childrenNodes>" << getChildrenNodesString(root)
+//             << "</childrenNodes>" << endl;
+//     if (root->type == nodeTypes::func || root->type == nodeTypes::dummy) {
+//       *outCUs << "\t\t<funcArguments>" << endl;
+//       for (auto ai : root->argumentsList) {
+//         *outCUs << "\t\t\t<arg type=\"" << xmlEscape(ai.type) << "\""
+//                 << " defLine=\"" << xmlEscape(ai.defLine) << "\">"
+//                 << xmlEscape(ai.name) << "</arg>" << endl;
+//       }
+//       *outCUs << "\t\t</funcArguments>" << endl;
 
-      string rlVals = "";
-      for (auto rl : root->returnLines) {
-        rlVals += dputil::decodeLID(rl) + ", ";
-      }
-      *outCUs << "\t\t<funcReturnLines>" << rlVals << "</funcReturnLines>"
-              << endl;
-    }
+//       string rlVals = "";
+//       for (auto rl : root->returnLines) {
+//         rlVals += dputil::decodeLID(rl) + ", ";
+//       }
+//       *outCUs << "\t\t<funcReturnLines>" << rlVals << "</funcReturnLines>"
+//               << endl;
+//     }
 
-    if (root->type == nodeTypes::cu) {
-      CU *cu = static_cast<CU *>(root);
-      *outCUs << "\t\t<BasicBlockID>" << cu->BBID << "</BasicBlockID>" << endl;
-      *outCUs << "\t\t<readDataSize>" << cu->readDataSize << "</readDataSize>"
-              << endl;
-      *outCUs << "\t\t<writeDataSize>" << cu->writeDataSize
-              << "</writeDataSize>" << endl;
+//     if (root->type == nodeTypes::cu) {
+//       CU *cu = static_cast<CU *>(root);
+//       *outCUs << "\t\t<BasicBlockID>" << cu->BBID << "</BasicBlockID>" << endl;
+//       *outCUs << "\t\t<readDataSize>" << cu->readDataSize << "</readDataSize>"
+//               << endl;
+//       *outCUs << "\t\t<writeDataSize>" << cu->writeDataSize
+//               << "</writeDataSize>" << endl;
 
-      *outCUs << "\t\t<instructionsCount>" << cu->instructionsCount
-              << "</instructionsCount>" << endl;
-      *outCUs << "\t\t<instructionLines count=\""
-              << (cu->instructionsLineNumbers).size() << "\">"
-              << getLineNumbersString(cu->instructionsLineNumbers)
-              << "</instructionLines>" << endl;
-      *outCUs << "\t\t<readPhaseLines count=\""
-              << (cu->readPhaseLineNumbers).size() << "\">"
-              << getLineNumbersString(cu->readPhaseLineNumbers)
-              << "</readPhaseLines>" << endl;
-      *outCUs << "\t\t<writePhaseLines count=\""
-              << (cu->writePhaseLineNumbers).size() << "\">"
-              << getLineNumbersString(cu->writePhaseLineNumbers)
-              << "</writePhaseLines>" << endl;
-      *outCUs << "\t\t<returnInstructions count=\""
-              << (cu->returnInstructions).size() << "\">"
-              << getLineNumbersString(cu->returnInstructions)
-              << "</returnInstructions>" << endl;
-      *outCUs << "\t\t<successors>" << endl;
-      for (auto sucCUi : cu->successorCUs) {
-        *outCUs << "\t\t\t<CU>" << sucCUi << "</CU>" << endl;
-      }
-      *outCUs << "\t\t</successors>" << endl;
+//       *outCUs << "\t\t<instructionsCount>" << cu->instructionsCount
+//               << "</instructionsCount>" << endl;
+//       *outCUs << "\t\t<instructionLines count=\""
+//               << (cu->instructionsLineNumbers).size() << "\">"
+//               << getLineNumbersString(cu->instructionsLineNumbers)
+//               << "</instructionLines>" << endl;
+//       *outCUs << "\t\t<readPhaseLines count=\""
+//               << (cu->readPhaseLineNumbers).size() << "\">"
+//               << getLineNumbersString(cu->readPhaseLineNumbers)
+//               << "</readPhaseLines>" << endl;
+//       *outCUs << "\t\t<writePhaseLines count=\""
+//               << (cu->writePhaseLineNumbers).size() << "\">"
+//               << getLineNumbersString(cu->writePhaseLineNumbers)
+//               << "</writePhaseLines>" << endl;
+//       *outCUs << "\t\t<returnInstructions count=\""
+//               << (cu->returnInstructions).size() << "\">"
+//               << getLineNumbersString(cu->returnInstructions)
+//               << "</returnInstructions>" << endl;
+//       *outCUs << "\t\t<successors>" << endl;
+//       for (auto sucCUi : cu->successorCUs) {
+//         *outCUs << "\t\t\t<CU>" << sucCUi << "</CU>" << endl;
+//       }
+//       *outCUs << "\t\t</successors>" << endl;
 
-      *outCUs << "\t\t<localVariables>" << endl;
-      for (auto lvi : cu->localVariableNames) {
-        *outCUs << "\t\t\t<local type=\"" << xmlEscape(lvi.type) << "\""
-                << " defLine=\"" << xmlEscape(lvi.defLine) << "\">"
-                << xmlEscape(lvi.name) << "</local>" << endl;
-      }
-      *outCUs << "\t\t</localVariables>" << endl;
+//       *outCUs << "\t\t<localVariables>" << endl;
+//       for (auto lvi : cu->localVariableNames) {
+//         *outCUs << "\t\t\t<local type=\"" << xmlEscape(lvi.type) << "\""
+//                 << " defLine=\"" << xmlEscape(lvi.defLine) << "\">"
+//                 << xmlEscape(lvi.name) << "</local>" << endl;
+//       }
+//       *outCUs << "\t\t</localVariables>" << endl;
 
-      *outCUs << "\t\t<globalVariables>" << endl;
-      for (auto gvi : cu->globalVariableNames) {
-        *outCUs << "\t\t\t<global type=\"" << xmlEscape(gvi.type) << "\""
-                << " defLine=\"" << xmlEscape(gvi.defLine) << "\">"
-                << xmlEscape(gvi.name) << "</global>" << endl;
-      }
-      *outCUs << "\t\t</globalVariables>" << endl;
+//       *outCUs << "\t\t<globalVariables>" << endl;
+//       for (auto gvi : cu->globalVariableNames) {
+//         *outCUs << "\t\t\t<global type=\"" << xmlEscape(gvi.type) << "\""
+//                 << " defLine=\"" << xmlEscape(gvi.defLine) << "\">"
+//                 << xmlEscape(gvi.name) << "</global>" << endl;
+//       }
+//       *outCUs << "\t\t</globalVariables>" << endl;
 
-      *outCUs << "\t\t<callsNode>" << endl;
-      for (auto i : (cu->callLineTofunctionMap)) {
-        for (auto ii : i.second) {
-          *outCUs << "\t\t\t<nodeCalled atLine=\"" << dputil::decodeLID(i.first)
-                  << "\">" << ii->ID << "</nodeCalled>" << endl;
-          // specifica for recursive fucntions inside loops. (Mo 5.11.2019)
-          *outCUs << "\t\t\t\t<recursiveFunctionCall>"
-                  << ii->recursiveFunctionCall << "</recursiveFunctionCall>"
-                  << endl;
-        }
-      }
-      *outCUs << "\t\t</callsNode>" << endl;
-    }
+//       *outCUs << "\t\t<callsNode>" << endl;
+//       for (auto i : (cu->callLineTofunctionMap)) {
+//         for (auto ii : i.second) {
+//           *outCUs << "\t\t\t<nodeCalled atLine=\"" << dputil::decodeLID(i.first)
+//                   << "\">" << ii->ID << "</nodeCalled>" << endl;
+//           // specifica for recursive fucntions inside loops. (Mo 5.11.2019)
+//           *outCUs << "\t\t\t\t<recursiveFunctionCall>"
+//                   << ii->recursiveFunctionCall << "</recursiveFunctionCall>"
+//                   << endl;
+//         }
+//       }
+//       *outCUs << "\t\t</callsNode>" << endl;
+//     }
 
-    *outCUs << "\t</Node>" << endl << endl;
-  }
-}
+//     *outCUs << "\t</Node>" << endl << endl;
+//   }
+// }
 
-void CUGeneration::closeOutputFiles() {
+// void CUGeneration::closeOutputFiles() {
 
-  if (outCUs != NULL && outCUs->is_open()) {
-    outCUs->flush();
-    outCUs->close();
-  }
+//   if (outCUs != NULL && outCUs->is_open()) {
+//     outCUs->flush();
+//     outCUs->close();
+//   }
 
-  if (outOriginalVariables != NULL && outOriginalVariables->is_open()) {
-    outOriginalVariables->flush();
-    outOriginalVariables->close();
-  }
-  // delete outCUs;
-}
+//   if (outOriginalVariables != NULL && outOriginalVariables->is_open()) {
+//     outOriginalVariables->flush();
+//     outOriginalVariables->close();
+//   }
+//   // delete outCUs;
+// }
 /*********************************** End of output functions
  * **************************************/
 
@@ -772,607 +772,607 @@ void CUGeneration::closeOutputFiles() {
 //     return varName;
 // }
 
-// recieves the region and outputs all variables and variables crossing basic
-// block boundaries in the region.
-void CUGeneration::populateGlobalVariablesSet(Region *TopRegion,
-                                              set<string> &globalVariablesSet) {
+// // recieves the region and outputs all variables and variables crossing basic
+// // block boundaries in the region.
+// void CUGeneration::populateGlobalVariablesSet(Region *TopRegion,
+//                                               set<string> &globalVariablesSet) {
 
-  map<string, BasicBlock *> variableToBBMap;
-  bool isGlobalVariable;
-  for (Region::block_iterator bb = TopRegion->block_begin();
-       bb != TopRegion->block_end(); ++bb) {
-    for (BasicBlock::iterator instruction = (*bb)->begin();
-         instruction != (*bb)->end(); ++instruction) {
-      if (isa<LoadInst>(instruction) || isa<StoreInst>(instruction) ||
-          isa<CallInst>(instruction)) {
+//   map<string, BasicBlock *> variableToBBMap;
+//   bool isGlobalVariable;
+//   for (Region::block_iterator bb = TopRegion->block_begin();
+//        bb != TopRegion->block_end(); ++bb) {
+//     for (BasicBlock::iterator instruction = (*bb)->begin();
+//          instruction != (*bb)->end(); ++instruction) {
+//       if (isa<LoadInst>(instruction) || isa<StoreInst>(instruction) ||
+//           isa<CallInst>(instruction)) {
 
-        // string varName = refineVarName(determineVariableName(instruction,
-        // isGlobalVariable));
-        // NOTE: changed 'instruction' to '&*instruction'
-        string varName = determineVariableName(&*instruction, isGlobalVariable);
+//         // string varName = refineVarName(determineVariableName(instruction,
+//         // isGlobalVariable));
+//         // NOTE: changed 'instruction' to '&*instruction'
+//         string varName = determineVariableName(&*instruction, isGlobalVariable);
 
-        if (isGlobalVariable) // add it if it is a global variable in the
-                              // program
-        {
-          programGlobalVariablesSet.insert(varName);
-        }
+//         if (isGlobalVariable) // add it if it is a global variable in the
+//                               // program
+//         {
+//           programGlobalVariablesSet.insert(varName);
+//         }
 
-        if (variableToBBMap.find(varName) != variableToBBMap.end()) {
-          // this var has already once recordded. check for bb id
-          if (variableToBBMap[varName] != *bb) {
-            // global variable found. Insert into the globalVariablesSet
-            globalVariablesSet.insert(varName);
-          }
-        } else {
-          // record usage of the variable.
-          variableToBBMap.insert(pair<string, BasicBlock *>(varName, *bb));
-          // errs() << varName << "\n";
-        }
-      }
-    }
-  }
-}
+//         if (variableToBBMap.find(varName) != variableToBBMap.end()) {
+//           // this var has already once recordded. check for bb id
+//           if (variableToBBMap[varName] != *bb) {
+//             // global variable found. Insert into the globalVariablesSet
+//             globalVariablesSet.insert(varName);
+//           }
+//         } else {
+//           // record usage of the variable.
+//           variableToBBMap.insert(pair<string, BasicBlock *>(varName, *bb));
+//           // errs() << varName << "\n";
+//         }
+//       }
+//     }
+//   }
+// }
 
-void CUGeneration::createCUs(Region *TopRegion, set<string> &globalVariablesSet,
-                             vector<CU *> &CUVector,
-                             map<string, vector<CU *>> &BBIDToCUIDsMap,
-                             Node *root, LoopInfo &LI) {
-  // NOTE: changed 'ThisModule->getDataLayout()' to
-  // '&ThisModule->getDataLayout()'
-  const DataLayout *DL =
-      &ThisModule->getDataLayout(); // used to get data size of variables,
-                                    // pointers, structs etc.
-  Node *currentNode = root;
-  CU *cu;
-  int lid;
-  string varName;
-  string varType;
-  set<string> suspiciousVariables;
-  string basicBlockName;
+// void CUGeneration::createCUs(Region *TopRegion, set<string> &globalVariablesSet,
+//                              vector<CU *> &CUVector,
+//                              map<string, vector<CU *>> &BBIDToCUIDsMap,
+//                              Node *root, LoopInfo &LI) {
+//   // NOTE: changed 'ThisModule->getDataLayout()' to
+//   // '&ThisModule->getDataLayout()'
+//   const DataLayout *DL =
+//       &ThisModule->getDataLayout(); // used to get data size of variables,
+//                                     // pointers, structs etc.
+//   Node *currentNode = root;
+//   CU *cu;
+//   int lid;
+//   string varName;
+//   string varType;
+//   set<string> suspiciousVariables;
+//   string basicBlockName;
 
-  map<Loop *, Node *> loopToNodeMap;
+//   map<Loop *, Node *> loopToNodeMap;
 
-  for (Region::block_iterator bb = TopRegion->block_begin();
-       bb != TopRegion->block_end(); ++bb) {
+//   for (Region::block_iterator bb = TopRegion->block_begin();
+//        bb != TopRegion->block_end(); ++bb) {
 
-    // Get the closest loop where bb lives in.
-    // (loop == NULL) if bb is not in any loop.
-    Loop *loop = LI.getLoopFor(*bb);
-    if (loop) {
-      // if bb is in a loop and if we have already created a node for that loop,
-      // assign it to currentNode.
-      if (loopToNodeMap.find(loop) != loopToNodeMap.end()) {
-        currentNode = loopToNodeMap[loop];
-        // errs() << "))))) " << dputil::decodeLID(currentNode->startLine) << "
-        // " << dputil::decodeLID(currentNode->endLine) << "\n";
-      }
-      // else, create a new Node for the loop, add it as children of currentNode
-      // and add it to the map.
-      else {
-        if (bb->getName().size() != 0) {
-          // errs() << "Name: " << bb->getName() << "\n";
-        }
+//     // Get the closest loop where bb lives in.
+//     // (loop == NULL) if bb is not in any loop.
+//     Loop *loop = LI.getLoopFor(*bb);
+//     if (loop) {
+//       // if bb is in a loop and if we have already created a node for that loop,
+//       // assign it to currentNode.
+//       if (loopToNodeMap.find(loop) != loopToNodeMap.end()) {
+//         currentNode = loopToNodeMap[loop];
+//         // errs() << "))))) " << dputil::decodeLID(currentNode->startLine) << "
+//         // " << dputil::decodeLID(currentNode->endLine) << "\n";
+//       }
+//       // else, create a new Node for the loop, add it as children of currentNode
+//       // and add it to the map.
+//       else {
+//         if (bb->getName().size() != 0) {
+//           // errs() << "Name: " << bb->getName() << "\n";
+//         }
 
-        Node *n = new Node;
-        n->type = nodeTypes::loop;
-        n->parentNode = currentNode;
-        currentNode->childrenNodes.push_back(n);
+//         Node *n = new Node;
+//         n->type = nodeTypes::loop;
+//         n->parentNode = currentNode;
+//         currentNode->childrenNodes.push_back(n);
 
-        loopToNodeMap[loop] = n;
-        currentNode = n;
-        // errs() << "--bb->Name: " << bb->getName() << " , " << "node->ID: " <<
-        // currentNode->ID << "\n";
-      }
-    } else {
-      if (bb->getName().size() != 0) {
-        // errs() << "bb Name: " << bb->getName() << "\n";
-      }
-      // end of loops. go to the parent of the loop. may have to jump several
-      // nodes in case of nested loops
-      for (map<Loop *, Node *>::iterator it = loopToNodeMap.begin();
-           it != loopToNodeMap.end(); it++)
-        if (it->second ==
-            currentNode) // current node found in loop map jump to its parent.
-        {
-          currentNode = currentNode->parentNode;
-          it = loopToNodeMap
-                   .begin(); // search the whole map again for current node
-          if (it->second == currentNode) // due to it++ we need to check first
-                                         // element of map ourself
-            currentNode = currentNode->parentNode;
-        }
-    }
+//         loopToNodeMap[loop] = n;
+//         currentNode = n;
+//         // errs() << "--bb->Name: " << bb->getName() << " , " << "node->ID: " <<
+//         // currentNode->ID << "\n";
+//       }
+//     } else {
+//       if (bb->getName().size() != 0) {
+//         // errs() << "bb Name: " << bb->getName() << "\n";
+//       }
+//       // end of loops. go to the parent of the loop. may have to jump several
+//       // nodes in case of nested loops
+//       for (map<Loop *, Node *>::iterator it = loopToNodeMap.begin();
+//            it != loopToNodeMap.end(); it++)
+//         if (it->second ==
+//             currentNode) // current node found in loop map jump to its parent.
+//         {
+//           currentNode = currentNode->parentNode;
+//           it = loopToNodeMap
+//                    .begin(); // search the whole map again for current node
+//           if (it->second == currentNode) // due to it++ we need to check first
+//                                          // element of map ourself
+//             currentNode = currentNode->parentNode;
+//         }
+//     }
 
-    cu = new CU;
+//     cu = new CU;
 
-    // errs() << "==== " << bb->getName() << "\n"; //"cu->ID: " << cu->ID << " ,
-    // " << "node->ID: " << currentNode->ID << " , " << "tmpNode->ID: " <<
-    // tmpNode->ID << " , " << "bb->Name: " << bb->getName() << "\n";
+//     // errs() << "==== " << bb->getName() << "\n"; //"cu->ID: " << cu->ID << " ,
+//     // " << "node->ID: " << currentNode->ID << " , " << "tmpNode->ID: " <<
+//     // tmpNode->ID << " , " << "bb->Name: " << bb->getName() << "\n";
 
-    if (bb->getName().size() == 0)
-      bb->setName(cu->ID);
+//     if (bb->getName().size() == 0)
+//       bb->setName(cu->ID);
 
-    cu->BBID = bb->getName();
-    cu->BB = *bb; // Mohammad 23.12.2020
-    currentNode->childrenNodes.push_back(cu);
-    vector<CU *> basicBlockCUVector;
-    basicBlockCUVector.push_back(cu);
-    BBIDToCUIDsMap.insert(
-        pair<string, vector<CU *>>(bb->getName(), basicBlockCUVector));
+//     cu->BBID = bb->getName();
+//     cu->BB = *bb; // Mohammad 23.12.2020
+//     currentNode->childrenNodes.push_back(cu);
+//     vector<CU *> basicBlockCUVector;
+//     basicBlockCUVector.push_back(cu);
+//     BBIDToCUIDsMap.insert(
+//         pair<string, vector<CU *>>(bb->getName(), basicBlockCUVector));
 
-    for (BasicBlock::iterator instruction = (*bb)->begin();
-         instruction != (*bb)->end(); ++instruction) {
-      // NOTE: 'instruction' --> '&*instruction'
-      lid = getLID(&*instruction, fileID);
-      basicBlockName = bb->getName();
-      if (lid > 0) {
-        cu->instructionsLineNumbers.insert(lid);
-        cu->instructionsCount++;
-        // find return instructions
-        if (isa<ReturnInst>(instruction)) {
-          cu->returnInstructions.insert(lid);
-        }
-        // find branches to return instructions, i.e. return statements
-        // Lukas 21.09.20
-        else if (isa<BranchInst>(instruction)) {
-          if ((cast<BranchInst>(instruction))->isUnconditional()) {
-            if ((cast<BranchInst>(instruction))->getNumSuccessors() == 1) {
-              BasicBlock *successorBB =
-                  (cast<BranchInst>(instruction))->getSuccessor(0);
-              for (BasicBlock::iterator innerInstruction = successorBB->begin();
-                   innerInstruction != successorBB->end(); ++innerInstruction) {
-                if (isa<ReturnInst>(innerInstruction)) {
-                  cu->returnInstructions.insert(lid);
-                  break;
-                }
-              }
-            }
-          }
-        }
-        //}
-        if (isa<StoreInst>(instruction)) {
+//     for (BasicBlock::iterator instruction = (*bb)->begin();
+//          instruction != (*bb)->end(); ++instruction) {
+//       // NOTE: 'instruction' --> '&*instruction'
+//       lid = getLID(&*instruction, fileID);
+//       basicBlockName = bb->getName();
+//       if (lid > 0) {
+//         cu->instructionsLineNumbers.insert(lid);
+//         cu->instructionsCount++;
+//         // find return instructions
+//         if (isa<ReturnInst>(instruction)) {
+//           cu->returnInstructions.insert(lid);
+//         }
+//         // find branches to return instructions, i.e. return statements
+//         // Lukas 21.09.20
+//         else if (isa<BranchInst>(instruction)) {
+//           if ((cast<BranchInst>(instruction))->isUnconditional()) {
+//             if ((cast<BranchInst>(instruction))->getNumSuccessors() == 1) {
+//               BasicBlock *successorBB =
+//                   (cast<BranchInst>(instruction))->getSuccessor(0);
+//               for (BasicBlock::iterator innerInstruction = successorBB->begin();
+//                    innerInstruction != successorBB->end(); ++innerInstruction) {
+//                 if (isa<ReturnInst>(innerInstruction)) {
+//                   cu->returnInstructions.insert(lid);
+//                   break;
+//                 }
+//               }
+//             }
+//           }
+//         }
+//         //}
+//         if (isa<StoreInst>(instruction)) {
 
-          // get size of data written into memory by this store instruction
-          Value *operand = instruction->getOperand(1);
-          Type *Ty = operand->getType();
-          unsigned u = DL->getTypeSizeInBits(Ty);
-          cu->writeDataSize += u;
-          // varName = refineVarName(determineVariableName(instruction));
-          varName = determineVariableName(&*instruction);
-          varType = determineVariableType(&*instruction);
-          // if(globalVariablesSet.count(varName) ||
-          // programGlobalVariablesSet.count(varName))
-          {
-            suspiciousVariables.insert(varName);
-            if (lid > 0)
-              cu->writePhaseLineNumbers.insert(lid);
-          }
-        } else if (isa<LoadInst>(instruction)) {
+//           // get size of data written into memory by this store instruction
+//           Value *operand = instruction->getOperand(1);
+//           Type *Ty = operand->getType();
+//           unsigned u = DL->getTypeSizeInBits(Ty);
+//           cu->writeDataSize += u;
+//           // varName = refineVarName(determineVariableName(instruction));
+//           varName = determineVariableName(&*instruction);
+//           varType = determineVariableType(&*instruction);
+//           // if(globalVariablesSet.count(varName) ||
+//           // programGlobalVariablesSet.count(varName))
+//           {
+//             suspiciousVariables.insert(varName);
+//             if (lid > 0)
+//               cu->writePhaseLineNumbers.insert(lid);
+//           }
+//         } else if (isa<LoadInst>(instruction)) {
 
-          // get size of data read from memory by this load instruction
-          Type *Ty = instruction->getType();
-          unsigned u = DL->getTypeSizeInBits(Ty);
-          cu->readDataSize += u;
-          // varName = refineVarName(determineVariableName(instruction));
-          varName = determineVariableName(&*instruction);
-          if (suspiciousVariables.count(varName)) {
-            // VIOLATION OF CAUTIOUS PROPERTY
-            // it is a load instruction which read the value of a global
-            // variable.
-            // This global variable has already been stored previously.
-            // A new CU should be created here.
-            cu->readPhaseLineNumbers.erase(lid);
-            cu->writePhaseLineNumbers.erase(lid);
-            cu->instructionsLineNumbers.erase(lid);
-            cu->instructionsCount--;
-            if (cu->instructionsLineNumbers.empty()) {
-              cu->removeCU();
-              cu->startLine = -1;
-              cu->endLine = -1;
-            } else {
-              cu->startLine = *(cu->instructionsLineNumbers.begin());
-              cu->endLine = *(cu->instructionsLineNumbers.rbegin());
-            }
-            cu->basicBlockName = basicBlockName;
-            CUVector.push_back(cu);
-            suspiciousVariables.clear();
-            CU *temp =
-                cu; // keep current CU to make a reference to the successor CU
-            cu = new CU;
+//           // get size of data read from memory by this load instruction
+//           Type *Ty = instruction->getType();
+//           unsigned u = DL->getTypeSizeInBits(Ty);
+//           cu->readDataSize += u;
+//           // varName = refineVarName(determineVariableName(instruction));
+//           varName = determineVariableName(&*instruction);
+//           if (suspiciousVariables.count(varName)) {
+//             // VIOLATION OF CAUTIOUS PROPERTY
+//             // it is a load instruction which read the value of a global
+//             // variable.
+//             // This global variable has already been stored previously.
+//             // A new CU should be created here.
+//             cu->readPhaseLineNumbers.erase(lid);
+//             cu->writePhaseLineNumbers.erase(lid);
+//             cu->instructionsLineNumbers.erase(lid);
+//             cu->instructionsCount--;
+//             if (cu->instructionsLineNumbers.empty()) {
+//               cu->removeCU();
+//               cu->startLine = -1;
+//               cu->endLine = -1;
+//             } else {
+//               cu->startLine = *(cu->instructionsLineNumbers.begin());
+//               cu->endLine = *(cu->instructionsLineNumbers.rbegin());
+//             }
+//             cu->basicBlockName = basicBlockName;
+//             CUVector.push_back(cu);
+//             suspiciousVariables.clear();
+//             CU *temp =
+//                 cu; // keep current CU to make a reference to the successor CU
+//             cu = new CU;
 
-            cu->BBID = bb->getName();
-            cu->BB = *bb; // Mohammad 23.12.2020
-            // errs() << "bb->Name: "  << bb->getName() << " , " << "cu->ID: "
-            // << cu->ID  << " , " << "node->ID: " << currentNode->ID << "\n";
+//             cu->BBID = bb->getName();
+//             cu->BB = *bb; // Mohammad 23.12.2020
+//             // errs() << "bb->Name: "  << bb->getName() << " , " << "cu->ID: "
+//             // << cu->ID  << " , " << "node->ID: " << currentNode->ID << "\n";
 
-            currentNode->childrenNodes.push_back(cu);
-            temp->successorCUs.push_back(cu->ID);
-            BBIDToCUIDsMap[bb->getName()].push_back(cu);
-            if (lid > 0) {
-              cu->readPhaseLineNumbers.insert(lid);
-              cu->instructionsLineNumbers.insert(lid);
-            }
-          } else {
-            if (globalVariablesSet.count(varName) ||
-                programGlobalVariablesSet.count(varName)) {
-              if (lid > 0)
-                cu->readPhaseLineNumbers.insert(lid);
-            }
-          }
-        }
-      }
-    }
-    if (cu->instructionsLineNumbers.empty()) {
-      cu->removeCU();
-      cu->startLine = -1;
-      cu->endLine = -1;
-    } else {
-      cu->startLine = *(cu->instructionsLineNumbers.begin());
-      cu->endLine = *(cu->instructionsLineNumbers.rbegin());
-    }
+//             currentNode->childrenNodes.push_back(cu);
+//             temp->successorCUs.push_back(cu->ID);
+//             BBIDToCUIDsMap[bb->getName()].push_back(cu);
+//             if (lid > 0) {
+//               cu->readPhaseLineNumbers.insert(lid);
+//               cu->instructionsLineNumbers.insert(lid);
+//             }
+//           } else {
+//             if (globalVariablesSet.count(varName) ||
+//                 programGlobalVariablesSet.count(varName)) {
+//               if (lid > 0)
+//                 cu->readPhaseLineNumbers.insert(lid);
+//             }
+//           }
+//         }
+//       }
+//     }
+//     if (cu->instructionsLineNumbers.empty()) {
+//       cu->removeCU();
+//       cu->startLine = -1;
+//       cu->endLine = -1;
+//     } else {
+//       cu->startLine = *(cu->instructionsLineNumbers.begin());
+//       cu->endLine = *(cu->instructionsLineNumbers.rbegin());
+//     }
 
-    cu->basicBlockName = basicBlockName;
-    CUVector.push_back(cu);
-    suspiciousVariables.clear();
+//     cu->basicBlockName = basicBlockName;
+//     CUVector.push_back(cu);
+//     suspiciousVariables.clear();
 
-    // check for call instructions in current basic block
-    for (BasicBlock::iterator instruction = (*bb)->begin();
-         instruction != (*bb)->end(); ++instruction) {
-      // Mohammad 6.7.2020: Don't create nodes for library functions (c++/llvm).
-      int32_t lid = getLID(&*instruction, fileID);
-      if (lid > 0) {
+//     // check for call instructions in current basic block
+//     for (BasicBlock::iterator instruction = (*bb)->begin();
+//          instruction != (*bb)->end(); ++instruction) {
+//       // Mohammad 6.7.2020: Don't create nodes for library functions (c++/llvm).
+//       int32_t lid = getLID(&*instruction, fileID);
+//       if (lid > 0) {
 
-        if (isa<CallInst>(instruction)) {
+//         if (isa<CallInst>(instruction)) {
 
-          Function *f = (cast<CallInst>(instruction))->getCalledFunction();
-          // TODO: DO the same for Invoke inst
+//           Function *f = (cast<CallInst>(instruction))->getCalledFunction();
+//           // TODO: DO the same for Invoke inst
 
-          // Mohammad 6.7.2020
-          Function::iterator FI = f->begin();
-          bool externalFunction = true;
-          string lid;
+//           // Mohammad 6.7.2020
+//           Function::iterator FI = f->begin();
+//           bool externalFunction = true;
+//           string lid;
 
-          for (Function::iterator FI = f->begin(), FE = f->end(); FI != FE;
-               ++FI) {
-            externalFunction = false;
-            auto tempBI = FI->begin();
-            if (DebugLoc dl = tempBI->getDebugLoc()) {
-              lid = to_string(dl->getLine());
-            } else {
-              if (tempBI->getFunction()->getSubprogram())
-                lid = to_string(
-                    tempBI->getFunction()->getSubprogram()->getLine());
-              else {
-                lid = "LineNotFound";
-              }
-            }
-            break;
-          }
-          if (externalFunction)
-            continue;
+//           for (Function::iterator FI = f->begin(), FE = f->end(); FI != FE;
+//                ++FI) {
+//             externalFunction = false;
+//             auto tempBI = FI->begin();
+//             if (DebugLoc dl = tempBI->getDebugLoc()) {
+//               lid = to_string(dl->getLine());
+//             } else {
+//               if (tempBI->getFunction()->getSubprogram())
+//                 lid = to_string(
+//                     tempBI->getFunction()->getSubprogram()->getLine());
+//               else {
+//                 lid = "LineNotFound";
+//               }
+//             }
+//             break;
+//           }
+//           if (externalFunction)
+//             continue;
 
-          Node *n = new Node;
-          n->type = nodeTypes::dummy;
-          // For ordinary function calls, F has a name.
-          // However, sometimes the function being called
-          // in IR is encapsulated by "bitcast()" due to
-          // the way of compiling and linking. In this way,
-          // getCalledFunction() method returns NULL.
-          // Also, getName() returns NULL if this is an indirect function call.
-          if (f) {
-            n->name = f->getName();
+//           Node *n = new Node;
+//           n->type = nodeTypes::dummy;
+//           // For ordinary function calls, F has a name.
+//           // However, sometimes the function being called
+//           // in IR is encapsulated by "bitcast()" due to
+//           // the way of compiling and linking. In this way,
+//           // getCalledFunction() method returns NULL.
+//           // Also, getName() returns NULL if this is an indirect function call.
+//           if (f) {
+//             n->name = f->getName();
 
-            // @Zia: This for loop appeared after the else part. For some
-            // function calls, the value of f is null. I guess that is why you
-            // have checked if f is not null here. Anyway, I (Mohammad) had to
-            // bring the for loop inside to avoid the segmentation fault. If you
-            // think it is not appropriate, find a solution for it. 14.2.2016
-            for (Function::arg_iterator it = f->arg_begin(); it != f->arg_end();
-                 it++) {
-              string type_str;
-              raw_string_ostream rso(type_str);
-              (it->getType())->print(rso);
-              Variable v(string(it->getName()), rso.str(), lid);
-              n->argumentsList.push_back(v);
-            }
-          } else // get name of the indirect function which is called
-          {
-            Value *v = (cast<CallInst>(instruction))->getCalledValue();
-            Value *sv = v->stripPointerCasts();
-            StringRef fname = sv->getName();
-            n->name = fname;
-          }
+//             // @Zia: This for loop appeared after the else part. For some
+//             // function calls, the value of f is null. I guess that is why you
+//             // have checked if f is not null here. Anyway, I (Mohammad) had to
+//             // bring the for loop inside to avoid the segmentation fault. If you
+//             // think it is not appropriate, find a solution for it. 14.2.2016
+//             for (Function::arg_iterator it = f->arg_begin(); it != f->arg_end();
+//                  it++) {
+//               string type_str;
+//               raw_string_ostream rso(type_str);
+//               (it->getType())->print(rso);
+//               Variable v(string(it->getName()), rso.str(), lid);
+//               n->argumentsList.push_back(v);
+//             }
+//           } else // get name of the indirect function which is called
+//           {
+//             Value *v = (cast<CallInst>(instruction))->getCalledValue();
+//             Value *sv = v->stripPointerCasts();
+//             StringRef fname = sv->getName();
+//             n->name = fname;
+//           }
 
-          // Recursive functions (Mo 5.11.2019)
-          CallGraphWrapperPass *CGWP = &(getAnalysis<CallGraphWrapperPass>());
-          if (isRecursive(*f, CGWP->getCallGraph())) {
-            int lid = getLID(&*instruction, fileID);
-            n->recursiveFunctionCall =
-                n->name + " " + dputil::decodeLID(lid) + ",";
-          }
+//           // Recursive functions (Mo 5.11.2019)
+//           CallGraphWrapperPass *CGWP = &(getAnalysis<CallGraphWrapperPass>());
+//           if (isRecursive(*f, CGWP->getCallGraph())) {
+//             int lid = getLID(&*instruction, fileID);
+//             n->recursiveFunctionCall =
+//                 n->name + " " + dputil::decodeLID(lid) + ",";
+//           }
 
-          vector<CU *> BBCUsVector = BBIDToCUIDsMap[bb->getName()];
-          // locate the CU where this function call belongs
-          for (auto i : BBCUsVector) {
-            int lid = getLID(&*instruction, fileID);
-            if (lid >= i->startLine && lid <= i->endLine) {
-              i->instructionsLineNumbers.insert(lid);
-              i->childrenNodes.push_back(n);
+//           vector<CU *> BBCUsVector = BBIDToCUIDsMap[bb->getName()];
+//           // locate the CU where this function call belongs
+//           for (auto i : BBCUsVector) {
+//             int lid = getLID(&*instruction, fileID);
+//             if (lid >= i->startLine && lid <= i->endLine) {
+//               i->instructionsLineNumbers.insert(lid);
+//               i->childrenNodes.push_back(n);
 
-              i->callLineTofunctionMap[lid].push_back(n);
-              break;
-            }
-          }
-        }
-      }
-    }
-  }
-}
+//               i->callLineTofunctionMap[lid].push_back(n);
+//               break;
+//             }
+//           }
+//         }
+//       }
+//     }
+//   }
+// }
 
-void CUGeneration::fillCUVariables(Region *TopRegion,
-                                   set<string> &globalVariablesSet,
-                                   vector<CU *> &CUVector,
-                                   map<string, vector<CU *>> &BBIDToCUIDsMap) {
-  int lid;
-  string varName, varType, varDefLine;
-  // Changed TerminatorInst to Instuction
-  const Instruction *TInst;
-  string successorBB;
+// void CUGeneration::fillCUVariables(Region *TopRegion,
+//                                    set<string> &globalVariablesSet,
+//                                    vector<CU *> &CUVector,
+//                                    map<string, vector<CU *>> &BBIDToCUIDsMap) {
+//   int lid;
+//   string varName, varType, varDefLine;
+//   // Changed TerminatorInst to Instuction
+//   const Instruction *TInst;
+//   string successorBB;
 
-  for (Region::block_iterator bb = TopRegion->block_begin();
-       bb != TopRegion->block_end(); ++bb) {
-    CU *lastCU = BBIDToCUIDsMap[bb->getName()]
-                     .back(); // get the last CU in the basic block
-    // get all successor basic blocks for bb
-    TInst = bb->getTerminator();
-    for (unsigned i = 0, nSucc = TInst->getNumSuccessors(); i < nSucc; ++i) {
-      // get the name of successor basicBlock
-      successorBB = TInst->getSuccessor(i)->getName();
-      // get the first CU of the successor basicBlock and record its ID in
-      // current CU's successorCUs
-      lastCU->successorCUs.push_back(BBIDToCUIDsMap[successorBB].front()->ID);
-    }
+//   for (Region::block_iterator bb = TopRegion->block_begin();
+//        bb != TopRegion->block_end(); ++bb) {
+//     CU *lastCU = BBIDToCUIDsMap[bb->getName()]
+//                      .back(); // get the last CU in the basic block
+//     // get all successor basic blocks for bb
+//     TInst = bb->getTerminator();
+//     for (unsigned i = 0, nSucc = TInst->getNumSuccessors(); i < nSucc; ++i) {
+//       // get the name of successor basicBlock
+//       successorBB = TInst->getSuccessor(i)->getName();
+//       // get the first CU of the successor basicBlock and record its ID in
+//       // current CU's successorCUs
+//       lastCU->successorCUs.push_back(BBIDToCUIDsMap[successorBB].front()->ID);
+//     }
 
-    auto bbCU = BBIDToCUIDsMap[bb->getName()].begin();
-    for (BasicBlock::iterator instruction = (*bb)->begin();
-         instruction != (*bb)->end(); ++instruction) {
-      if (isa<LoadInst>(instruction) || isa<StoreInst>(instruction)) {
-        // NOTE: changed 'instruction' to '&*instruction'
-        lid = getLID(&*instruction, fileID);
-        if (lid == 0)
-          continue;
-        // varName = refineVarName(determineVariableName(instruction));
-        // NOTE: changed 'instruction' to '&*instruction', next 2 lines
-        varName = determineVariableName(&*instruction);
-        varType = determineVariableType(&*instruction);
-        varDefLine = determineVariableDefLine(&*instruction);
+//     auto bbCU = BBIDToCUIDsMap[bb->getName()].begin();
+//     for (BasicBlock::iterator instruction = (*bb)->begin();
+//          instruction != (*bb)->end(); ++instruction) {
+//       if (isa<LoadInst>(instruction) || isa<StoreInst>(instruction)) {
+//         // NOTE: changed 'instruction' to '&*instruction'
+//         lid = getLID(&*instruction, fileID);
+//         if (lid == 0)
+//           continue;
+//         // varName = refineVarName(determineVariableName(instruction));
+//         // NOTE: changed 'instruction' to '&*instruction', next 2 lines
+//         varName = determineVariableName(&*instruction);
+//         varType = determineVariableType(&*instruction);
+//         varDefLine = determineVariableDefLine(&*instruction);
 
-        Variable v(varName, varType, varDefLine);
+//         Variable v(varName, varType, varDefLine);
 
-        std::string prefix("ARRAY");
-        if (!varType.compare(0, prefix.size(), prefix)) {
-          varName = "ARRAY, " + varName;
-        }
+//         std::string prefix("ARRAY");
+//         if (!varType.compare(0, prefix.size(), prefix)) {
+//           varName = "ARRAY, " + varName;
+//         }
 
-        // errs() << "Name: "  << varName << " " << "Type: " << varType << "\n";
+//         // errs() << "Name: "  << varName << " " << "Type: " << varType << "\n";
 
-        if (lid > (*bbCU)->endLine) {
-          bbCU = next(bbCU, 1);
-        }
-        if (globalVariablesSet.count(varName) ||
-            programGlobalVariablesSet.count(varName)) {
-          (*bbCU)->globalVariableNames.insert(v);
-          // originalVariablesSet.insert(varName);
-        } else {
-          (*bbCU)->localVariableNames.insert(v);
-          // originalVariablesSet.insert(varName);
-        }
-      }
-    }
-  }
-}
+//         if (lid > (*bbCU)->endLine) {
+//           bbCU = next(bbCU, 1);
+//         }
+//         if (globalVariablesSet.count(varName) ||
+//             programGlobalVariablesSet.count(varName)) {
+//           (*bbCU)->globalVariableNames.insert(v);
+//           // originalVariablesSet.insert(varName);
+//         } else {
+//           (*bbCU)->localVariableNames.insert(v);
+//           // originalVariablesSet.insert(varName);
+//         }
+//       }
+//     }
+//   }
+// }
 
-void CUGeneration::findStartEndLineNumbers(Node *root, int &start, int &end) {
-  if (root->type == nodeTypes::cu) {
-    if (start == -1 || start > root->startLine) {
-      start = root->startLine;
-    }
+// void CUGeneration::findStartEndLineNumbers(Node *root, int &start, int &end) {
+//   if (root->type == nodeTypes::cu) {
+//     if (start == -1 || start > root->startLine) {
+//       start = root->startLine;
+//     }
 
-    if (end < root->endLine) {
-      end = root->endLine;
-    }
-  }
+//     if (end < root->endLine) {
+//       end = root->endLine;
+//     }
+//   }
 
-  for (auto i : root->childrenNodes) {
-    findStartEndLineNumbers(i, start, end);
-  }
-}
+//   for (auto i : root->childrenNodes) {
+//     findStartEndLineNumbers(i, start, end);
+//   }
+// }
 
-void CUGeneration::fillStartEndLineNumbers(Node *root, LoopInfo &LI) {
-  if (root->type != nodeTypes::cu) {
-    int start = -1, end = -1;
+// void CUGeneration::fillStartEndLineNumbers(Node *root, LoopInfo &LI) {
+//   if (root->type != nodeTypes::cu) {
+//     int start = -1, end = -1;
 
-    if (root->type == nodeTypes::loop) {
-      for (auto i : root->childrenNodes) {
-        if (i->type == nodeTypes::cu) {
-          Loop *loop = LI.getLoopFor(i->BB);
-          DebugLoc dl = loop->getStartLoc();
-          int32_t lid = 0;
-          lid = (fileID << LIDSIZE) + dl->getLine();
-          loopStartLines[root->ID] = dputil::decodeLID(lid);
-          break;
-        }
-      }
-    }
-    findStartEndLineNumbers(root, start, end);
+//     if (root->type == nodeTypes::loop) {
+//       for (auto i : root->childrenNodes) {
+//         if (i->type == nodeTypes::cu) {
+//           Loop *loop = LI.getLoopFor(i->BB);
+//           DebugLoc dl = loop->getStartLoc();
+//           int32_t lid = 0;
+//           lid = (fileID << LIDSIZE) + dl->getLine();
+//           loopStartLines[root->ID] = dputil::decodeLID(lid);
+//           break;
+//         }
+//       }
+//     }
+//     findStartEndLineNumbers(root, start, end);
 
-    root->startLine = start;
-    root->endLine = end;
-  }
+//     root->startLine = start;
+//     root->endLine = end;
+//   }
 
-  for (auto i : root->childrenNodes) {
-    fillStartEndLineNumbers(i, LI);
-  }
-}
+//   for (auto i : root->childrenNodes) {
+//     fillStartEndLineNumbers(i, LI);
+//   }
+// }
 
-bool CUGeneration::doFinalization(Module &M) {
-  // write the current count of CUs to a file to avoid duplicate CUs.
-  if (outCUIDCounter && outCUIDCounter->is_open()) {
-    *outCUIDCounter << CUIDCounter;
-    outCUIDCounter->flush();
-    outCUIDCounter->close();
-  }
-  return true;
-}
+// bool CUGeneration::doFinalization(Module &M) {
+//   // write the current count of CUs to a file to avoid duplicate CUs.
+//   if (outCUIDCounter && outCUIDCounter->is_open()) {
+//     *outCUIDCounter << CUIDCounter;
+//     outCUIDCounter->flush();
+//     outCUIDCounter->close();
+//   }
+//   return true;
+// }
 
-bool CUGeneration::doInitialization(Module &M) {
+// void CUGeneration::initializeCUIDCounter() {
+//   std::string CUCounterFile = "DP_CUIDCounter.txt";
 
-  CUIDCounter = 0;
-  defaultIsGlobalVariableValue = false;
-  ThisModule = &M;
-  outCUIDCounter = NULL;
+//   if (dputil::fexists(CUCounterFile)) {
+//     std::fstream inCUIDCounter(CUCounterFile, std::ios_base::in);
+//     ;
+//     inCUIDCounter >> CUIDCounter;
+//     inCUIDCounter.close();
+//   }
+// }
 
-  initializeCUIDCounter();
+// bool CUGeneration::isRecursive(Function &F, CallGraph &CG) {
+//   auto callNode = CG[&F];
+//   for (unsigned i = 0; i < callNode->size(); i++) {
+//     if ((*callNode)[i]->getFunction() == &F)
+//       return true;
+//   }
 
-  for (Module::global_iterator I = ThisModule->global_begin();
-       I != ThisModule->global_end(); I++) {
-    Value *globalVariable = dyn_cast<Value>(I);
-    string glo = string(globalVariable->getName());
-    if (glo.find(".") == glo.npos) {
-      programGlobalVariablesSet.insert(glo);
-      // originalVariablesSet.insert(glo);
-    }
-  }
+//   return false;
+// }
 
-  return true;
-}
+// bool CUGeneration::doInitialization(Module &M) {
 
-void CUGeneration::initializeCUIDCounter() {
-  std::string CUCounterFile = "DP_CUIDCounter.txt";
+//   CUIDCounter = 0;
+//   defaultIsGlobalVariableValue = false;
+//   ThisModule = &M;
+//   outCUIDCounter = NULL;
 
-  if (dputil::fexists(CUCounterFile)) {
-    std::fstream inCUIDCounter(CUCounterFile, std::ios_base::in);
-    ;
-    inCUIDCounter >> CUIDCounter;
-    inCUIDCounter.close();
-  }
-}
+//   initializeCUIDCounter();
 
-bool CUGeneration::isRecursive(Function &F, CallGraph &CG) {
-  auto callNode = CG[&F];
-  for (unsigned i = 0; i < callNode->size(); i++) {
-    if ((*callNode)[i]->getFunction() == &F)
-      return true;
-  }
+//   for (Module::global_iterator I = ThisModule->global_begin();
+//        I != ThisModule->global_end(); I++) {
+//     Value *globalVariable = dyn_cast<Value>(I);
+//     string glo = string(globalVariable->getName());
+//     if (glo.find(".") == glo.npos) {
+//       programGlobalVariablesSet.insert(glo);
+//       // originalVariablesSet.insert(glo);
+//     }
+//   }
 
-  return false;
-}
+//   return true;
+// }
 
-void CUGeneration::getAnalysisUsage(AnalysisUsage &AU) const {
-  AU.addRequiredTransitive<RegionInfoPass>();
-  // NOTE: changed 'LoopInfo' to 'LoopInfoWrapperPass'
-  AU.addRequired<LoopInfoWrapperPass>();
-  AU.addPreserved<LoopInfoWrapperPass>();
-  // Get recursive functions called in loops. (Mo 5.11.2019)
-  AU.addRequired<CallGraphWrapperPass>();
-  AU.setPreservesAll();
-}
+// void CUGeneration::getAnalysisUsage(AnalysisUsage &AU) const {
+//   AU.addRequiredTransitive<RegionInfoPass>();
+//   // NOTE: changed 'LoopInfo' to 'LoopInfoWrapperPass'
+//   AU.addRequired<LoopInfoWrapperPass>();
+//   AU.addPreserved<LoopInfoWrapperPass>();
+//   // Get recursive functions called in loops. (Mo 5.11.2019)
+//   AU.addRequired<CallGraphWrapperPass>();
+//   AU.setPreservesAll();
+// }
 
 bool CUGeneration::runOnFunction(Function &F) {
-  StringRef funcName = F.getName();
+  // StringRef funcName = F.getName();
 
   // Avoid functions we don't want to analyze
-  if (funcName.find("llvm.") != string::npos) // llvm debug calls
-  {
-    return false;
-  }
-  if (funcName.find("__dp_") != string::npos) // instrumentation calls
-  {
-    return false;
-  }
-  if (funcName.find("__cx") != string::npos) // c++ init calls
-  {
-    return false;
-  }
-  if (funcName.find("__clang") != string::npos) // clang helper calls
-  {
-    return false;
-  }
-  if (funcName.find("_GLOBAL_") != string::npos) { // global init calls (c++)
-    return false;
-  }
-  if (funcName.find("pthread_") != string::npos) {
-    return false;
-  }
+  // if (funcName.find("llvm.") != string::npos) // llvm debug calls
+  // {
+  //   return false;
+  // }
+  // if (funcName.find("__dp_") != string::npos) // instrumentation calls
+  // {
+  //   return false;
+  // }
+  // if (funcName.find("__cx") != string::npos) // c++ init calls
+  // {
+  //   return false;
+  // }
+  // if (funcName.find("__clang") != string::npos) // clang helper calls
+  // {
+  //   return false;
+  // }
+  // if (funcName.find("_GLOBAL_") != string::npos) { // global init calls (c++)
+  //   return false;
+  // }
+  // if (funcName.find("pthread_") != string::npos) {
+  //   return false;
+  // }
 
   // initializeCUIDCounter();
-  vector<CU *> CUVector;
-  set<string> globalVariablesSet; // list of variables which appear in more than
-                                  // one basic block
-  map<string, vector<CU *>> BBIDToCUIDsMap;
+  // vector<CU *> CUVector;
+  // set<string> globalVariablesSet; // list of variables which appear in more than
+  //                                 // one basic block
+  // map<string, vector<CU *>> BBIDToCUIDsMap;
 
-  determineFileID(F, fileID);
-  /********************* Initialize root values ***************************/
-  Node *root = new Node;
-  root->name = F.getName();
-  root->type = nodeTypes::func;
+  // determineFileID(F, fileID);
+  // /********************* Initialize root values ***************************/
+  // Node *root = new Node;
+  // root->name = F.getName();
+  // root->type = nodeTypes::func;
 
-  // Get list of arguments for this function and store them in root.
-  // NOTE: changed the way we get the arguments
-  // for (Function::ArgumentListType::iterator it = F.getArgumentList().begin();
-  // it != F.getArgumentList().end(); it++) {
+  // // Get list of arguments for this function and store them in root.
+  // // NOTE: changed the way we get the arguments
+  // // for (Function::ArgumentListType::iterator it = F.getArgumentList().begin();
+  // // it != F.getArgumentList().end(); it++) {
 
-  BasicBlock *BB = &F.getEntryBlock();
-  auto BI = BB->begin();
-  string lid;
-  if (DebugLoc dl = BI->getDebugLoc()) {
-    lid = to_string(dl->getLine());
-  } else {
-    lid = to_string(BI->getFunction()->getSubprogram()->getLine());
-  }
+  // BasicBlock *BB = &F.getEntryBlock();
+  // auto BI = BB->begin();
+  // string lid;
+  // if (DebugLoc dl = BI->getDebugLoc()) {
+  //   lid = to_string(dl->getLine());
+  // } else {
+  //   lid = to_string(BI->getFunction()->getSubprogram()->getLine());
+  // }
 
-  for (Function::arg_iterator it = F.arg_begin(); it != F.arg_end(); it++) {
+  // for (Function::arg_iterator it = F.arg_begin(); it != F.arg_end(); it++) {
 
-    string type_str;
-    raw_string_ostream rso(type_str);
-    (it->getType())->print(rso);
+  //   string type_str;
+  //   raw_string_ostream rso(type_str);
+  //   (it->getType())->print(rso);
 
-    Variable v(it->getName(), rso.str(), to_string(fileID) + ":" + lid);
-    root->argumentsList.push_back(v);
-  }
-  /********************* End of initialize root values
-   * ***************************/
-  // errs()<< "000---\n";
-  // NOTE: changed the pass name for loopinfo -- LoopInfo &LI =
-  // getAnalysis<LoopInfo>();
-  LoopInfo &LI = getAnalysis<LoopInfoWrapperPass>().getLoopInfo();
+  //   Variable v(it->getName(), rso.str(), to_string(fileID) + ":" + lid);
+  //   root->argumentsList.push_back(v);
+  // }
+  // /********************* End of initialize root values
+  //  * ***************************/
+  // // errs()<< "000---\n";
+  // // NOTE: changed the pass name for loopinfo -- LoopInfo &LI =
+  // // getAnalysis<LoopInfo>();
+  // LoopInfo &LI = getAnalysis<LoopInfoWrapperPass>().getLoopInfo();
 
-  // get the top level region
-  RIpass = &getAnalysis<RegionInfoPass>();
-  RI = &(RIpass->getRegionInfo());
-  Region *TopRegion = RI->getTopLevelRegion();
+  // // get the top level region
+  // RIpass = &getAnalysis<RegionInfoPass>();
+  // RI = &(RIpass->getRegionInfo());
+  // Region *TopRegion = RI->getTopLevelRegion();
 
-  getFunctionReturnLines(TopRegion, root);
+  // getFunctionReturnLines(TopRegion, root);
 
-  populateGlobalVariablesSet(TopRegion, globalVariablesSet);
+  // populateGlobalVariablesSet(TopRegion, globalVariablesSet);
 
-  createCUs(TopRegion, globalVariablesSet, CUVector, BBIDToCUIDsMap, root, LI);
+  // createCUs(TopRegion, globalVariablesSet, CUVector, BBIDToCUIDsMap, root, LI);
 
-  fillCUVariables(TopRegion, globalVariablesSet, CUVector, BBIDToCUIDsMap);
+  // fillCUVariables(TopRegion, globalVariablesSet, CUVector, BBIDToCUIDsMap);
 
-  fillStartEndLineNumbers(root, LI);
+  // fillStartEndLineNumbers(root, LI);
 
-  secureStream();
+  // secureStream();
 
-  // printOriginalVariables(originalVariablesSet);
+  // // printOriginalVariables(originalVariablesSet);
 
-  printData(root);
+  // printData(root);
 
-  for (auto i : CUVector) {
-    delete (i);
-  }
+  // for (auto i : CUVector) {
+  //   delete (i);
+  // }
 
-  return false;
+  // return false;
 }
 
 char CUGeneration::ID = 0;
