+++ DiscoPoP/CMakeLists.txt
@@ -2,7 +2,7 @@ include_directories(
     ../share/include
     )
 
-add_llvm_library(LLVMDPInstrumentation MODULE
-	DPInstrumentation.cpp
+add_llvm_library(LLVMDiscoPoP MODULE
+	DiscoPoP.cpp
 	../share/lib/DPUtils.cpp
     )
